app1_comfyui/streamlit/requirements.txt
streamlit>=1.35.0
requests>=2.31.0
websocket-client>=1.8.0
Pillow>=10.3.0
